requirements.txt
streamlit== 1.42.0 
tensorflow==2.18.0
numpy<2
pillow==10.4.0
pandas==2.2.3
opencv-python-headless==4.8.1.78
